filter all URLs through the 'url' builtin filter
