+++ _includes/layouts/base.njk
@@ -38,6 +38,5 @@
     {% endif %}
 
     <script src="{{ '/assets/uswds/js/uswds.min.js' | url }}"></script>
-    <script src="{{ '/assets/js/site.js' | url }}"></script>
   </body>
 </html>
