+++ _includes/layouts/offering.njk
@@ -21,7 +21,7 @@ breadcrumb:
   {% feature
     "Interested in other ways GSA can help your workplace strategy?",
     "Explore offerings",
-    "offerings/",
-    "assets/img/offering-bottom-feature.png"
+    "/offerings/",
+    "/assets/img/offering-bottom-feature.png"
   %}
 </div>
