+++ pages/offerings/index.njk
@@ -39,8 +39,8 @@ hero:
     {% feature
       "Wondering about the latest workplace trends and analysis?",
       "Explore insights",
-      "insights/",
-      "assets/img/offerings-feature.jpg"
+      "/insights/",
+      "/assets/img/offerings-feature.jpg"
     %}
   </div>
